Removed timetable class and migrate this concept to classe, renamed timeentry to weekdirectory
