+++ src/main/java/online/vidacademica/services/config/TestConfig.java
@@ -25,8 +25,7 @@ import online.vidacademica.services.entities.State;
 import online.vidacademica.services.entities.Subject;
 import online.vidacademica.services.entities.Test;
 import online.vidacademica.services.entities.TestResult;
-import online.vidacademica.services.entities.TimeTable;
-import online.vidacademica.services.entities.TimeTableEntry;
+import online.vidacademica.services.entities.WeekEntry;
 import online.vidacademica.services.entities.User;
 import online.vidacademica.services.repositories.AddressRepository;
 import online.vidacademica.services.repositories.CityRepository;
@@ -40,8 +39,7 @@ import online.vidacademica.services.repositories.StateRepository;
 import online.vidacademica.services.repositories.SubjectRepository;
 import online.vidacademica.services.repositories.TestRepository;
 import online.vidacademica.services.repositories.TestResultRepository;
-import online.vidacademica.services.repositories.TimeTableEntryRepository;
-import online.vidacademica.services.repositories.TimeTableRepository;
+import online.vidacademica.services.repositories.WeekEntryRepository;
 import online.vidacademica.services.repositories.UserRepository;
 
 @Configuration
@@ -80,42 +78,39 @@ public class TestConfig implements CommandLineRunner {
     private SubjectRepository subjectRepository;
 
     @Autowired
-    private TimeTableEntryRepository timeTableEntryRepository;
-
-    @Autowired
-    private TimeTableRepository timeTableRepository;
+    private WeekEntryRepository timeTableEntryRepository;
 
     @Autowired
     private TestResultRepository testResultRepository;
 
     @Autowired
     private TestRepository testRepository;
-    
+
     @Autowired
     private RoleRepository roleRepository;
 
     @Autowired
     private BCryptPasswordEncoder passwordEncode;
-    
+
     @Override
     public void run(String... args) throws Exception {
 
         // ADICIONA OS ELEMENTOS DO BANCO TESTE
 
-        User u1 = new User(null, "Carlos Gustavo", "cgletras@gmail.com", null, "A2", "ASD",passwordEncode.encode("123"), Instant.now());
+        User u1 = new User(null, "Carlos Gustavo", "cgletras@gmail.com", null, "A2", "ASD", passwordEncode.encode("123"), Instant.now());
         User u2 = new User(null, "Eduardo Augusto", "dudu@gmail.com", null, "A3", "ABD", passwordEncode.encode("123"), Instant.now());
         User u3 = new User(null, "Rafael Sotero", "soso@gmail.com", null, "B5", "CSD", passwordEncode.encode("123"), Instant.now());
         User u4 = new User(null, "Tiago Marques", "titi@gmail.com", null, "C3", "JSD", passwordEncode.encode("123"), Instant.now());
         User u5 = new User(null, "Nelio Alves", "nelio@gmail.com", null, "d4", "JSD", passwordEncode.encode("123"), Instant.now());
         User u6 = new User(null, "Cricia", "cricia@gmail.com", null, "d4", "JSD", passwordEncode.encode("123"), Instant.now());
-        
-        userRepository.saveAll(Arrays.asList(u1, u2, u3, u4,u5,u6));
-        
+
+        userRepository.saveAll(Arrays.asList(u1, u2, u3, u4, u5, u6));
+
         Role role_Teacher = new Role(null, "ROLE_TEACHER");
         Role role_Student = new Role(null, "ROLE_STUDENT");
-        
-        roleRepository.saveAll(Arrays.asList(role_Teacher,role_Student));
-        
+
+        roleRepository.saveAll(Arrays.asList(role_Teacher, role_Student));
+
         u1.getRoles().add(role_Student);
         u2.getRoles().add(role_Student);
         u3.getRoles().add(role_Student);
@@ -123,8 +118,8 @@ public class TestConfig implements CommandLineRunner {
         u5.getRoles().add(role_Teacher);
         u6.getRoles().add(role_Teacher);
 
-        userRepository.saveAll(Arrays.asList(u1, u2, u3, u4,u5,u6));
- 
+        userRepository.saveAll(Arrays.asList(u1, u2, u3, u4, u5, u6));
+
 
         Country ct1 = new Country(null, "Brasil");
         countryRepository.saveAll(Arrays.asList(ct1));
@@ -169,11 +164,11 @@ public class TestConfig implements CommandLineRunner {
                 "" + "Conhecer fundamentos e boas práticas dos aspectos essenciais de projeto e\n"
                         + "desenvolvimento de um sistema para Internet com orientação a objetos e acesso a banco de\n"
                         + "dados. Utilizar técnicas e ferramentas para implementação dos tópicos estudados.",
-                105.0, true, Instant.now(), 60.0,LocalDate.of(2019,02,01),LocalDate.of(2019,07,12));
+                105.0, true, Instant.now(), 60.0, LocalDate.of(2019, 02, 01), LocalDate.of(2019, 07, 12));
 
         Subject sub2 = new Subject(null, "Banco de Dados",
                 "Conhecer fundamentos e boas práticas usando a linguagem Mysql", 75.0, true, Instant.now(), 60.0,
-                LocalDate.of(2019,02,01),LocalDate.of(2019,07,12));
+                LocalDate.of(2019, 02, 01), LocalDate.of(2019, 07, 12));
 
         subjectRepository.saveAll(Arrays.asList(sub1));
         subjectRepository.saveAll(Arrays.asList(sub2));
@@ -188,27 +183,26 @@ public class TestConfig implements CommandLineRunner {
 
         courseRepository.saveAll(Arrays.asList(course1));
 
-        Classe class1 = new Classe(null, "SI -2018/01", Instant.parse("2018-01-01T00:21:22Z"),
-                Instant.parse("2018-01-01T00:21:22Z"), true, Instant.parse("2019-07-22T15:21:22Z"));
+        Classe class1 = new Classe(null, "SI -2018/01", LocalDate.of(2018, 1, 1),
+                LocalDate.of(2019, 7, 22), true, Instant.parse("2018-01-01T00:21:22Z"));
 
-        Classe class2 = new Classe(null, "LCI - 2018/01", Instant.parse("2018-01-01T00:21:22Z"),
-                Instant.parse("2018-01-01T00:21:22Z"), true, Instant.parse("2019-07-22T15:21:22Z"));
-
+        Classe class2 = new Classe(null, "LCI - 2018/01", LocalDate.of(2018, 1, 1),
+                LocalDate.of(2019, 7, 22), true, Instant.parse("2018-01-01T00:21:22Z"));
 
         postRepository.saveAll(Arrays.asList(p1, p2, p3, p4));
         phoneRepository.saveAll(Arrays.asList(ph1, ph2, ph3));
         classRepository.saveAll(Arrays.asList(class1));
         classRepository.saveAll(Arrays.asList(class2));
 
-        TimeTable t1 = new TimeTable(null, "Horários 2019-2", LocalDate.now(), LocalDate.now().plusDays(180));
+//        TimeTable t1 = new TimeTable(null, "Horários 2019-2", LocalDate.now(), LocalDate.now().plusDays(180));
 
-        TimeTableEntry e1 = new TimeTableEntry(null, DayOfWeek.MONDAY, 18L, 45L, 19L, 30L, t1);
-        TimeTableEntry e2 = new TimeTableEntry(null, DayOfWeek.MONDAY, 19L, 30L, 20L, 15L, t1);
-        TimeTableEntry e3 = new TimeTableEntry(null, DayOfWeek.TUESDAY, 18L, 45L, 19L, 30L, t1);
-        TimeTableEntry e4 = new TimeTableEntry(null, DayOfWeek.TUESDAY, 19L, 30L, 20L, 15L, t1);
-        TimeTableEntry e5 = new TimeTableEntry(null, DayOfWeek.TUESDAY, 20L, 15L, 21L, 00L, t1);
+        WeekEntry e1 = new WeekEntry(null, DayOfWeek.MONDAY, 18L, 45L, 19L, 30L, class1);
+        WeekEntry e2 = new WeekEntry(null, DayOfWeek.MONDAY, 19L, 30L, 20L, 15L, class1);
+        WeekEntry e3 = new WeekEntry(null, DayOfWeek.TUESDAY, 18L, 45L, 19L, 30L, class2);
+        WeekEntry e4 = new WeekEntry(null, DayOfWeek.TUESDAY, 19L, 30L, 20L, 15L, class2);
+        WeekEntry e5 = new WeekEntry(null, DayOfWeek.TUESDAY, 20L, 15L, 21L, 00L, class2);
 
-        timeTableRepository.save(t1);
+//        timeTableRepository.save(t1);
         timeTableEntryRepository.saveAll(Arrays.asList(e1, e2, e3, e4, e5));
 
         Test test1 = new Test(null, "Exercicio Avaliativo Banco de Dados", 10.00, Instant.parse("2019-08-15T00:21:22Z"),
@@ -269,7 +263,7 @@ public class TestConfig implements CommandLineRunner {
         TestResult testResult14 = new TestResult(u2, test5, 17.00, Instant.parse("2019-10-11T00:21:22Z"));
         TestResult testResult15 = new TestResult(u3, test5, 17.00, Instant.parse("2019-10-11T00:21:22Z"));
         TestResult testResult16 = new TestResult(u4, test5, 20.00, Instant.parse("2019-10-11T00:21:22Z"));
-        
+
         testResultRepository.saveAll(Arrays.asList(testResult1));
         testResultRepository.saveAll(Arrays.asList(testResult2));
         testResultRepository.saveAll(Arrays.asList(testResult3));
