+++ src/main/java/online/vidacademica/services/dto/TimeBoxDTO.java
@@ -1,42 +1,35 @@
 package online.vidacademica.services.dto;
 
-import online.vidacademica.services.entities.TimeBox;
-
 import java.io.Serializable;
 import java.time.Instant;
 
 public class TimeBoxDTO implements Serializable {
-	private static final long serialVersionUID = 1L;
-	
-	private Instant start;
-	private Instant end;
+    private static final long serialVersionUID = 1L;
 
-	public TimeBoxDTO() {
-	}
+    private Instant start;
+    private Instant end;
 
-	public TimeBoxDTO(Instant start, Instant end) {
-		this.start = start;
-		this.end = end;
-	}
-	
-	public TimeBoxDTO(TimeBox entity) {
-		this.start = entity.getStart();
-		this.end = entity.getEnd();
-	}
+    public TimeBoxDTO() {
+    }
 
-	public Instant getStart() {
-		return start;
-	}
+    public TimeBoxDTO(Instant start, Instant end) {
+        this.start = start;
+        this.end = end;
+    }
 
-	public void setStart(Instant start) {
-		this.start = start;
-	}
+    public Instant getStart() {
+        return start;
+    }
 
-	public Instant getEnd() {
-		return end;
-	}
+    public void setStart(Instant start) {
+        this.start = start;
+    }
 
-	public void setEnd(Instant end) {
-		this.end = end;
-	}
+    public Instant getEnd() {
+        return end;
+    }
+
+    public void setEnd(Instant end) {
+        this.end = end;
+    }
 }
