--- src/main/java/online/vidacademica/services/entities/TimeBox.java
@@ -1,86 +1,0 @@
-package online.vidacademica.services.entities;
-
-import java.io.Serializable;
-import java.time.Instant;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "tb_time_box")
-public class TimeBox implements Serializable {
-	private static final long serialVersionUID = 1L;
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	private Instant start;
-	private Instant end;
-
-	public TimeBox() {
-	}
-
-	public TimeBox(Long id, Instant start, Instant end) {
-		this.id = id;
-		this.start = start;
-		this.end = end;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public Instant getStart() {
-		return start;
-	}
-
-	public void setStart(Instant start) {
-		this.start = start;
-	}
-
-	public Instant getEnd() {
-		return end;
-	}
-
-	public void setEnd(Instant end) {
-		this.end = end;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((end == null) ? 0 : end.hashCode());
-		result = prime * result + ((start == null) ? 0 : start.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		TimeBox other = (TimeBox) obj;
-		if (end == null) {
-			if (other.end != null)
-				return false;
-		} else if (!end.equals(other.end))
-			return false;
-		if (start == null) {
-			if (other.start != null)
-				return false;
-		} else if (!start.equals(other.start))
-			return false;
-		return true;
-	}
-}
