--- src/main/java/online/vidacademica/services/repositories/TimeBoxRepository.java
@@ -1,8 +1,0 @@
-package online.vidacademica.services.repositories;
-
-import online.vidacademica.services.entities.TimeBox;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TimeBoxRepository extends JpaRepository<TimeBox, Long> {
-
-}
