--- src/main/java/online/vidacademica/services/repositories/TimeTableRepository.java
@@ -1,8 +1,0 @@
-package online.vidacademica.services.repositories;
-
-import online.vidacademica.services.entities.TimeTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TimeTableRepository extends JpaRepository<TimeTable, Long> {
-
-}
