+++ src/main/java/online/vidacademica/services/repositories/WeekEntryRepository.java
@@ -1,0 +1,8 @@
+package online.vidacademica.services.repositories;
+
+import online.vidacademica.services.entities.WeekEntry;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WeekEntryRepository extends JpaRepository<WeekEntry, Long> {
+
+}
